Added bond between players in test level and created prefab
Noticed that the test level did not have the bond between both players as it was created from a branch before the bond was implemented. Bond has now been re-implemented between the players in this level.

Prefab has also been created from the bonded players. Contains all the necessary scripts and game objects required for the player controlled characters. Will now allow for simple drag-and-drop addition to other levels.