Controller input now supported
Support for Xbox One and PS4 controllers integrated.

controllerTest scene shows this support. Player joins by pressing any button. Once joined, player can move around using the analogue stick or d-pad of the controller, or the WASD keys of a keyboard.

Player can also jump using the A/X button of the gamepad, or the spacebar of the keyboard