Grip Mechanic Re-Implemented
Re-implemented the grip mechanic after merge conflicts arising with previous branch.

Grip function currently mapped to the right trigger/R2 button on the controller, and the left control button on the keyboard.

Currently, when player activates the grip function. Player remains in position, the rigid body becomes kinematic, and forces are removed from the rigid body. This allows the other player to swing.

When the player released the grip button, the rigid body becomes dynamic, allowing the player to fall to the ground again.

A "m_canMove" boolean has also been created and implemented to prevent the player from moving when they are gripping.

Currently, players can grip the air and stay in place. Next implementation will be to fix this to only grip parts of the level, when the level is within a certain range of the player.