Improvements made to Level One
Improved level one. Player is first introduced to platforming, with 3 safe jumps that will not cause the player to die. After these jumps, player can die by falling in a pit.

Also added a couple of safe pits in the cavern to allow the players to get to grips with the swinging mechanic. If the players fall in these pits, they can climb back up to the left of the pit and try again. Beyond these pits, it is possible for the player to die.